refactor: move container components under containers folder
