+++ src/library-authoring/components/index.scss
@@ -1,2 +1,1 @@
 @import "./BaseCard.scss";
-@import "./ContainerCard.scss";
