+++ src/library-authoring/containers/ContainerCard.scss
