+++ src/library-authoring/containers/ContainerDeleter.tsx
