+++ src/library-authoring/containers/index.scss
@@ -1,0 +1,1 @@
+@import "./ContainerCard.scss";
