+++ src/library-authoring/index.scss
@@ -1,5 +1,6 @@
 @import "./component-info/ComponentPreview";
 @import "./components";
+@import "./containers";
 @import "./generic";
 @import "./LibraryAuthoringPage";
 @import "./units";
